Not_found_Exception

+++ src/home/home.service.ts
@@ -157,11 +157,22 @@ export class HomeService {
   }
 
   async deleteHomeById(id: number) {
+    const home = await this.prismaService.home.findUnique({
+      where: {
+        id,
+      },
+    });
+
+    if (!home) {
+      throw new NotFoundException();
+    }
+
     await this.prismaService.image.deleteMany({
       where: {
         home_id: id,
       },
     });
+
     await this.prismaService.home.delete({
       where: {
         id,
